training/python-app-rag++/pyproject.toml
# poetry install

[tool.poetry]
name = "training-app"
version = "0.1.0"
description = "Training apps"
authors = ["Your Name <you@example.com>"]

packages = [
  { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
faiss-cpu = "^1.7.4"
rank-bm25 = "^0.2.2"
anthropic = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.25.0"

[tool.poetry.scripts]
elevator = "app.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
